requirements.txt
gspread==6.2.0
google_auth==2.39.0
google_auth_oauthlib==1.2.1